refactor:style.ts for login
